requirements.txt
torch>=1.7.0
alibi-detect>=0.9.1
umap-learn
sentence-transformers
statsmodels
seaborn
datasets
scipy
tqdm
jupyter
numpy<=1.21
